remote script: split into reloadable handlers.py + thin __init__ shell
all commands move to handlers.py as cmd_<name>(params) methods with a
MAIN_THREAD_COMMANDS set; __init__.py keeps the socket server and
main-thread scheduling and gains a 'reload' command that importlib.reloads
handlers.py, so iterating no longer requires restarting Live.

drop the two dispatch entries (get_browser_categories, get_browser_items)
whose handlers never existed. scripts/install.sh copies + reloads;
replaces the sudo app-bundle installer.
